requirements.txt
ultralytics
opencv-python
streamlit
numpy
Pillow

